+++ packages/iov-jsonrpc/src/index.ts
@@ -7,9 +7,17 @@ export {
   isJsonCompatibleDictionary,
   isJsonCompatibleValue,
 } from "./jsoncompatibledictionary";
-export { parseJsonRpcId, parseJsonRpcRequest, parseJsonRpcResponse, parseJsonRpcError } from "./parse";
+export {
+  parseJsonRpcId,
+  parseJsonRpcRequest,
+  parseJsonRpcResponse,
+  parseJsonRpcError,
+  parseJsonRpcErrorResponse,
+  parseJsonRpcSuccessResponse,
+} from "./parse";
 export {
   isJsonRpcErrorResponse,
+  isJsonRpcSuccessResponse,
   JsonRpcRequest,
   JsonRpcErrorResponse,
   JsonRpcResponse,