+++ packages/iov-jsonrpc/src/parse.ts
@@ -72,10 +72,37 @@ function parseError(error: JsonCompatibleDictionary): JsonRpcError {
   return {
     code: error.code,
     message: error.message,
-    data: maybeUndefinedData,
+    ...(maybeUndefinedData !== undefined ? { data: maybeUndefinedData } : {}),
   };
 }
 
+/** Throws if data is not a JsonRpcErrorResponse */
+export function parseJsonRpcErrorResponse(data: unknown): JsonRpcErrorResponse {
+  if (!isJsonCompatibleDictionary(data)) {
+    throw new Error("Data must be JSON compatible dictionary");
+  }
+
+  if (data.jsonrpc !== "2.0") {
+    throw new Error(`Got unexpected jsonrpc version: ${JSON.stringify(data)}`);
+  }
+
+  const id = data.id;
+  if (typeof id !== "number" && id !== null) {
+    throw new Error("Invalid id field");
+  }
+
+  if (typeof data.error === "undefined" || !isJsonCompatibleDictionary(data.error)) {
+    throw new Error("Invalid error field");
+  }
+
+  return {
+    jsonrpc: "2.0",
+    id: id,
+    error: parseError(data.error),
+  };
+}
+
+/** @deprecated use parseJsonRpcErrorResponse */
 export function parseJsonRpcError(data: unknown): JsonRpcErrorResponse | undefined {
   if (!isJsonCompatibleDictionary(data)) {
     throw new Error("Data must be JSON compatible dictionary");
@@ -105,7 +132,8 @@ export function parseJsonRpcError(data: unknown): JsonRpcErrorResponse | undefin
   };
 }
 
-export function parseJsonRpcResponse(data: unknown): JsonRpcSuccessResponse {
+/** Throws if data is not a JsonRpcSuccessResponse */
+export function parseJsonRpcSuccessResponse(data: unknown): JsonRpcSuccessResponse {
   if (!isJsonCompatibleDictionary(data)) {
     throw new Error("Data must be JSON compatible dictionary");
   }
@@ -119,6 +147,10 @@ export function parseJsonRpcResponse(data: unknown): JsonRpcSuccessResponse {
     throw new Error("Invalid id field");
   }
 
+  if (typeof data.result === "undefined") {
+    throw new Error("Invalid result field");
+  }
+
   const result = data.result;
 
   return {
@@ -126,4 +158,9 @@ export function parseJsonRpcResponse(data: unknown): JsonRpcSuccessResponse {
     id: id,
     result: result,
   };
+}
+
+/** @deprecated use parseJsonRpcSuccessResponse */
+export function parseJsonRpcResponse(data: unknown): JsonRpcSuccessResponse {
+  return parseJsonRpcSuccessResponse(data);
 }