+++ packages/iov-jsonrpc/types/index.d.ts
@@ -1,4 +1,4 @@
 export { JsonRpcClient, SimpleMessagingConnection } from "./jsonrpcclient";
 export { JsonCompatibleArray, JsonCompatibleDictionary, JsonCompatibleValue, isJsonCompatibleArray, isJsonCompatibleDictionary, isJsonCompatibleValue, } from "./jsoncompatibledictionary";
-export { parseJsonRpcId, parseJsonRpcRequest, parseJsonRpcResponse, parseJsonRpcError } from "./parse";
-export { isJsonRpcErrorResponse, JsonRpcRequest, JsonRpcErrorResponse, JsonRpcResponse, JsonRpcSuccessResponse, jsonRpcCode, jsonRpcCodeInternalError, jsonRpcCodeInvalidParams, jsonRpcCodeInvalidRequest, jsonRpcCodeMethodNotFound, jsonRpcCodeParseError, jsonRpcCodeServerErrorDefault, } from "./types";
+export { parseJsonRpcId, parseJsonRpcRequest, parseJsonRpcResponse, parseJsonRpcError, parseJsonRpcErrorResponse, parseJsonRpcSuccessResponse, } from "./parse";
+export { isJsonRpcErrorResponse, isJsonRpcSuccessResponse, JsonRpcRequest, JsonRpcErrorResponse, JsonRpcResponse, JsonRpcSuccessResponse, jsonRpcCode, jsonRpcCodeInternalError, jsonRpcCodeInvalidParams, jsonRpcCodeInvalidRequest, jsonRpcCodeMethodNotFound, jsonRpcCodeParseError, jsonRpcCodeServerErrorDefault, } from "./types";